third_party/blink/web_tests/fast/lists/w3-css3-list-styles-deprecated.html

<!DOCTYPE HTML>
<html>
<head>
<title>Test deprecated list style types</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
</head>
<body>
<script>
    test(() => assert_true(CSS.supports("list-style-type", "decimal")));

    test(() => assert_true(CSS.supports("list-style-type", "binary")));
    test(() => assert_true(CSS.supports("list-style-type", "lower-hexadecimal")));
    test(() => assert_true(CSS.supports("list-style-type", "octal")));
    test(() => assert_true(CSS.supports("list-style-type", "upper-hexadecimal")));
    test(() => assert_true(CSS.supports("list-style-type", "afar")));
    test(() => assert_true(CSS.supports("list-style-type", "ethiopic-halehame-aa-et")));
    test(() => assert_true(CSS.supports("list-style-type", "ethiopic-halehame-aa-er")));
    test(() => assert_true(CSS.supports("list-style-type", "amharic")));
    test(() => assert_true(CSS.supports("list-style-type", "ethiopic-halehame-am-et")));
    test(() => assert_true(CSS.supports("list-style-type", "ethiopic-halehame-gez")));
    test(() => assert_true(CSS.supports("list-style-type", "ethiopic-abegede")));
    test(() => assert_true(CSS.supports("list-style-type", "ethiopic-abegede-gez")));
    test(() => assert_true(CSS.supports("list-style-type", "amharic-abegede")));
    test(() => assert_true(CSS.supports("list-style-type", "ethiopic-abegede-am-et")));
    test(() => assert_true(CSS.supports("list-style-type", "oromo")));
    test(() => assert_true(CSS.supports("list-style-type", "ethiopic-halehame-om-et")));
    test(() => assert_true(CSS.supports("list-style-type", "sidama")));
    test(() => assert_true(CSS.supports("list-style-type", "ethiopic-halehame-sid-et")));
    test(() => assert_true(CSS.supports("list-style-type", "somali")));
    test(() => assert_true(CSS.supports("list-style-type", "ethiopic-halehame-so-et")));
    test(() => assert_true(CSS.supports("list-style-type", "tigre")));
    test(() => assert_true(CSS.supports("list-style-type", "ethiopic-halehame-tig")));
    test(() => assert_true(CSS.supports("list-style-type", "tigrinya-er")));
    test(() => assert_true(CSS.supports("list-style-type", "tigrinya-er-abegede")));
    test(() => assert_true(CSS.supports("list-style-type", "ethiopic-abegede-ti-er")));
    test(() => assert_true(CSS.supports("list-style-type", "tigrinya-et")));
    test(() => assert_true(CSS.supports("list-style-type", "tigrinya-et-abegede")));
    test(() => assert_true(CSS.supports("list-style-type", "ethiopic-abegede-ti-et")));
    test(() => assert_true(CSS.supports("list-style-type", "lower-norwegian")));
    test(() => assert_true(CSS.supports("list-style-type", "upper-greek")));
    test(() => assert_true(CSS.supports("list-style-type", "upper-norwegian")));
    test(() => assert_true(CSS.supports("list-style-type", "asterisks")));
    test(() => assert_true(CSS.supports("list-style-type", "footnotes")));
</script>
</head>
</html>
